+++ src/renderer/containers/Account/SendPointsModal/index.tsx
@@ -5,14 +5,14 @@ import {useSelector} from 'react-redux';
 import * as Yup from 'yup';
 import axios from 'axios';
 
-import {FormButton, FormInput, FormSelectDetailed} from '@renderer/components/FormComponents';
+import {FormButton} from '@renderer/components/FormComponents';
 import Icon, {IconType} from '@renderer/components/Icon';
 import Modal from '@renderer/components/Modal';
-import RequiredAsterisk from '@renderer/components/RequiredAsterisk';
 import {getManagedAccounts} from '@renderer/selectors';
 import {Tx} from '@renderer/types';
 import {generateBlock, getKeyPairFromSigningKeyHex} from '@renderer/utils/signing';
 
+import SendPointsModalFields, {MATCH_ERROR} from './SendPointsModalFields';
 import './SendPointsModal.scss';
 
 const initialValues = {
@@ -24,13 +24,13 @@ const initialValues = {
 const validationSchema = Yup.object().shape({
   fromAccount: Yup.string()
     .required('This field is required')
-    .test('accounts-match', 'Sender and recipient can not match', function (value) {
+    .test('accounts-match', MATCH_ERROR, function (value) {
       return value !== this.parent.toAccount;
     }),
   points: Yup.number().moreThan(0, 'Must be greater than 0').required('This field is required'),
   toAccount: Yup.string()
     .required('This field is required')
-    .test('accounts-match', 'Sender and recipient can not match', function (value) {
+    .test('accounts-match', MATCH_ERROR, function (value) {
       return value !== this.parent.fromAccount;
     }),
 });
@@ -72,20 +72,6 @@ const SendPointsModal: FC<ComponentProps> = ({close}) => {
     console.error(response);
   };
 
-  const getFromOptions = () => {
-    return Object.values(managedAccounts).map(({account_number, nickname}) => ({
-      label: nickname,
-      value: account_number,
-    }));
-  };
-
-  const getToOptions = () => {
-    return Object.values(managedAccounts).map(({account_number, nickname}) => ({
-      label: nickname,
-      value: account_number,
-    }));
-  };
-
   const handleSubmit = ({points}: FormValues): void => {
     console.log(points);
   };
@@ -113,59 +99,7 @@ const SendPointsModal: FC<ComponentProps> = ({close}) => {
       onSubmit={handleSubmit}
       validationSchema={validationSchema}
     >
-      <FormSelectDetailed
-        className="SendPointsModal__select"
-        required
-        label="From"
-        options={getFromOptions()}
-        name="fromAccount"
-      />
-      <FormSelectDetailed
-        className="SendPointsModal__select"
-        required
-        label="To"
-        options={getToOptions()}
-        name="toAccount"
-      />
-      <table className="SendPointsModal__table">
-        <tbody>
-          <tr>
-            <td>Account Balance</td>
-            <td>
-              <span className="SendPointsModal__account-balance">0.00</span>
-            </td>
-          </tr>
-          <tr>
-            <td>
-              Points
-              <RequiredAsterisk />
-            </td>
-            <td>
-              <FormInput
-                className="SendPointsModal__points-input"
-                hideError
-                name="points"
-                placeholder="0.00"
-                type="number"
-              />
-            </td>
-          </tr>
-          <tr>
-            <td>Bank Registration Fee</td>
-            <td>0.01</td>
-          </tr>
-          <tr>
-            <td>Validator Tx Fee</td>
-            <td>0.02</td>
-          </tr>
-          <tr className="SendPointsModal__total-tr">
-            <td>Total</td>
-            <td>
-              <b>0.00</b>
-            </td>
-          </tr>
-        </tbody>
-      </table>
+      <SendPointsModalFields managedAccounts={managedAccounts} />
     </Modal>
   );
 };
